[FIRRTL] Move port directions and names to `firrtl.instance`
The primary contribution of this change is to add port directions and
names to the instance op.  This fixes the issue where we had to fetch
this information from the referenced module, which was not guaranteed to
be thread safe.

The printed form of the instance op is now a hybrid between the HW
instance op and a FIRRTL module, printing the entire "module signature"
at the end. This is quite verbose but needs to be in the IR format (i.e.
every op has to be parsable using only local information).

```mlir
firrtl.instance test @simple(in a: !firrtl.uint<1>, out b: !firrtl.uint<1>)
```