[FIRRTL] try adding a subaccess cache to the fir parser
This is an experiement to add an implicit CSE for subindexes and
subaccess to the FIRRTL parser.  Due to the nature of the FIRRTL textual
format, we create an incredible number of these things which slows down
the initial IR generation.  On my laptop this sped up a test case from
13-15 second to 8-9 seconds.