[FIRParser] Expand and remove partial connects at parse time
This change starts resolving partial connects at firrtl parse time to
remove them from the IR, with the eventual goal of removing partial
connects entirely.

There are places in the code where partial connects were used to as a
way to truncate sources in a connect.  A helper was added to
automatically insert pads or truncates where needed so that the sizes of
operands match for connect ops.  Most of these places could generate a
strict connect instead of a regular connect, but I decided to save the
test churn for a follow up commit.

The way this code is written has an eye towards eventually making strict
connect ops the *only* connect operations in the IR.  When creating a
potential bulk connect, if two connected fields have different sizes,
the connect will be expanded and the source field will be truncated or
extended.  If regular connects are staying in the IR then we can satisfy
this with a single bulk connect.