[FIRRTL] Don't invert the condition on cover statements
The FIRRTL parser tries to interpret assert and cover statments as
emitted by chisel and encode them in dedicated operations.  A part of
this is translating `when (x) $error` as `assert(!x)`.  It mistakenly
inverted the condition on cover statements as well, which do not need
the condition to be inverted.