[FIRRTL][IMDCE] Remove some cached state
This change removes some cached information related to how module ports
relate to instance results.

`resultPortToInstanceResultMapping`: This tracks a mapping from a block
argument to all instance results that it corresponds to.  We can instead
use the instance graph to go from a module to all instances of that
module, and grab the instance result we are interested in.  There is a
small difference in that instead of visiting only marked instances, we
now visit every instance.  This is a NFC but it could be slower if there
are many unreachable modules in the hierarchy with dead instances, but
that is very uncommon and currently illegal in the FIRRTL spec.

`lazyLiveInputPorts`: This tracks which ports of an instance should be
marked live if the instance op should be marked live. This corresponds
to the input ports of a module, so it is not worth tracking, especially
per-instance.  Instead we just mark the input ports of the instance
when the instance is marked alive.