[CI] Use a consistent key for ccache caching in Short Integration Tests
The default ccache action uses the current time as part of the key for caching.
The main problem with this is that it can cause too many different cache
entries, and can eject entries we want to be ejected.  There isn't a reason to
keep more than one copy of this cache around, so using a consistent key will
cause it to eject itself.  I put the name of the job in the key so that we can
identify which entries belong to what jobs when we examine the cache.  This
change also updates to a newer version of the ccache action.